dashboard Policies. User rights

+++ resources/views/admin/apartment-types/index.blade.php
@@ -46,17 +46,31 @@
                                         <td class="id">{{ $apartment_type->id }}</td>
                                         <td class="full_name">{{ $apartment_type->title }}</td>
                                         <td class="actions">
-                                            <div class="actions-btn">
-                                                <a class="btn p-0" href="{{route('apartment-types.edit', $apartment_type->id)}}">
-                                                    <span class="far fa-edit"></span>
-                                                </a>
+                                            @if (Auth::user()->hasRole('super_user'))
+                                                <div class="actions-btn">
+                                                    <a class="btn p-0" href="{{route('apartment-types.edit', $apartment_type->id)}}">
+                                                        <span class="far fa-edit"></span>
+                                                    </a>
 
-                                                {{Form::open(['route'=> ['apartment-types.destroy', $apartment_type->id], 'method' => 'delete'])}}
-                                                <button class="btn p-0 delete-btn" type="submit">
-                                                    <span class="far fa-trash-alt"></span>
-                                                </button>
-                                                {{Form::close()}}
-                                            </div>
+                                                    {{Form::open(['route'=> ['apartment-types.destroy', $apartment_type->id], 'method' => 'delete'])}}
+                                                    <button class="btn p-0 delete-btn" type="submit">
+                                                        <span class="far fa-trash-alt"></span>
+                                                    </button>
+                                                    {{Form::close()}}
+                                                </div>
+                                            @elseif (Auth::user()->id == $apartment_type->user_id)
+                                                <div class="actions-btn">
+                                                    <a class="btn p-0" href="{{route('apartment-types.edit', $apartment_type->id)}}">
+                                                        <span class="far fa-edit"></span>
+                                                    </a>
+
+                                                    {{Form::open(['route'=> ['apartment-types.destroy', $apartment_type->id], 'method' => 'delete'])}}
+                                                    <button class="btn p-0 delete-btn" type="submit">
+                                                        <span class="far fa-trash-alt"></span>
+                                                    </button>
+                                                    {{Form::close()}}
+                                                </div>
+                                            @endif
                                         </td>
                                     </tr>
                                 @endforeach
