+++ resources/views/admin/bed-types/index.blade.php
@@ -45,17 +45,31 @@
                                             <td class="id">{{ $bed_type->id }}</td>
                                             <td class="title">{{ $bed_type->title }}</td>
                                             <td class="actions">
-                                                <div class="actions-btn">
-                                                    <a class="btn p-0" href="{{route('bed-types.edit', $bed_type->id)}}">
-                                                        <span class="far fa-edit"></span>
-                                                    </a>
+                                                @if (Auth::user()->hasRole('super_user'))
+                                                    <div class="actions-btn">
+                                                        <a class="btn p-0" href="{{route('bed-types.edit', $bed_type->id)}}">
+                                                            <span class="far fa-edit"></span>
+                                                        </a>
 
-                                                    {{Form::open(['route'=> ['bed-types.destroy', $bed_type->id], 'method' => 'delete'])}}
+                                                        {{Form::open(['route'=> ['bed-types.destroy', $bed_type->id], 'method' => 'delete'])}}
                                                         <button class="btn p-0 delete-btn" type="submit">
                                                             <span class="far fa-trash-alt"></span>
                                                         </button>
-                                                    {{Form::close()}}
-                                                </div>
+                                                        {{Form::close()}}
+                                                    </div>
+                                                @elseif (Auth::user()->id == $bed_type->user_id)
+                                                    <div class="actions-btn">
+                                                        <a class="btn p-0" href="{{route('bed-types.edit', $bed_type->id)}}">
+                                                            <span class="far fa-edit"></span>
+                                                        </a>
+
+                                                        {{Form::open(['route'=> ['bed-types.destroy', $bed_type->id], 'method' => 'delete'])}}
+                                                        <button class="btn p-0 delete-btn" type="submit">
+                                                            <span class="far fa-trash-alt"></span>
+                                                        </button>
+                                                        {{Form::close()}}
+                                                    </div>
+                                                @endif
                                             </td>
                                         </tr>
                                     @endforeach
